+++ src/.internal/charts/stock/indicators/BollingerBands.ts
@@ -201,4 +201,20 @@ export class BollingerBands extends MovingAverage {
 		this.lowerBandSeries.dispose();
 		super._dispose();
 	}
+
+	public async hide(duration?: number): Promise<any> {
+		return Promise.all([
+			super.hide(duration),
+			this.upperBandSeries.hide(duration),
+			this.lowerBandSeries.hide(duration)
+		]);
+	}
+
+	public async show(duration?: number): Promise<any> {
+		return Promise.all([
+			super.show(duration),
+			this.upperBandSeries.show(duration),
+			this.lowerBandSeries.show(duration)
+		]);
+	}
 }